Move site description off of CRA's default text
This syncs the site's description with what's in the app's
header. That way hovering over any links will render a more
useful description of the site rather than the generic
"created via CRA" text.

Ideally this should be DRY'd up a bit, but that'd require
ejecting which would add more complexity than desirable
for such a small tweak.